refactor: split WidgetBridge storage/assets registrar
Move the storage, synchronous asset loading, and font asset native JS registrations out of the monolithic WidgetBridge registrar into storage_assets_api.cpp.

The call sites stay in their original positions so native API registration order remains unchanged.

Version-Bump: sdk=skip reason="mechanical internal registrar split; no SDK or behavior change"
Compat-Update: skip prefix=canvas2d reason="native registrar split only"
Compat-Update: skip prefix=css reason="native registrar split only"
Compat-Update: skip prefix=html reason="native registrar split only"
Compat-Update: skip prefix=imports reason="native registrar split only"
Compat-Update: skip prefix=react reason="native registrar split only"
Compat-Update: skip prefix=rn reason="native registrar split only"
Compat-Update: skip prefix=yoga reason="native registrar split only"
